Moved FineType classes to own package
