+++ application/core/src/main/java/com/github/pieter_groenendijk/service/fine/type/FineTypeRegistry.java
@@ -1,4 +1,4 @@
-package com.github.pieter_groenendijk.service.fine;
+package com.github.pieter_groenendijk.service.fine.type;
 
 import com.github.pieter_groenendijk.model.fine.FineType;
 import com.github.pieter_groenendijk.repository.fine.IFineRepository;
